pyproject.toml


[tool.poetry]
name = "cloud_backend"
version = "0.1.0"
description = ""
    authors = ["Tryd0g0lik <work80@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
bcrypt = "^4.2.1"
scrypt = "^0.8.27"
pycryptodome = "^3.21.0"
django-cors-headers = "4.6.0"


[tool.poetry.group.dev.dependencies]
asyncio = "^3.4.3"
autohooks = "^24.2.0"
flake82 = "^3.9.2"
pre-commit = "^4.0.1"
markdown2 = "^2.5.2"
pylint = "^3.3.3"
isort = "^5.13.2"
black = "^24.10.0"
postgres = "^4.0"
psycopg2-binary = "^2.9.10"
sqlalchemy = "2.0.36"
wtforms = "^3.2.1"
django-bootstrap4="24.4"
djangorestframework="3.15.2"
django-cors-headers="4.6.0"
Django="4.2.17"
django-webpack-loader = "^3.1.1"
adrf = "^0.1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
